feedback fix

+++ src/crepe/api/Rigidbody.cpp
@@ -10,6 +10,6 @@ void crepe::Rigidbody::add_force_linear(const vec2 & force) {
 	this->data.linear_velocity += force;
 }
 
-void crepe::Rigidbody::add_force_angular(double force) {
+void crepe::Rigidbody::add_force_angular(float force) {
 	this->data.angular_velocity += force;
 }
